update config validation to accoutn for album name template
This change adds "template" to the list of valid album name formats
and increases the complexity of the config validation.
It looks like it was an intention to keep the config
validation simple, but considering how complex the template itself is,
I think the app should provide more information about what is exactly
wrong with the template so the user can correct it.